Adds bittensort client

+++ app/api/v1/schemas/tao.py
@@ -6,6 +6,6 @@ class TaoDividendsResponse(BaseModel):
     
     netuid: int = Field(..., description="The subnet ID")
     hotkey: str = Field(..., description="The hotkey (account ID or public key)")
-    dividend: int = Field(..., description="The dividend value")
+    dividend: float = Field(..., description="The dividend value")
     cached: bool = Field(default=True, description="Whether the response was served from cache")
     stake_tx_triggered: bool = Field(default=True, description="Whether a stake transaction was triggered")
